% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDictionaries.R
\name{createDictionaries}
\alias{createDictionaries}
\title{Create dictionaries based on NCI thesaurus}
\usage{
createDictionaries(thesaurus)
}
\arguments{
\item{thesaurus}{(processed) NCI thesaurus.}
}
\value{
A list containing four dictionaries:
\itemize{
\item \code{cancer_dict}: Entities of semantic type 'Neoplastic Process' .
\item \code{drug_dict}: Entities of semantic type 'Pharmacologic Substance',
'Biologically Active Substance', 'Clinical Drug', 'Steroid', 'Immunologic Factor',
and 'Therapeutic or Preventive Procedure'.
\item \code{gene_dict}: Entities of semantic type 'Gene or Genome'.
\item \code{alteration_dict}: Entities of semantic type 'Cell or Molecular Dysfunction'.
}
}
\description{
Uses the \code{createDictionary} function to create \code{quanteda} dictionaries of cancer types, drugs, genes and molecular alterations.\cr
Dictionary keys are NCI thesaurus codes, dictionary values are NCI thesaurus synonyms.\cr
}
