% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandAlterations.R
\name{expandAlterations}
\alias{expandAlterations}
\title{Expand Molecular Dysfunctions}
\usage{
expandAlterations(codes)
}
\arguments{
\item{codes}{One or more codes for NCI thesaurus entities of semantic type \verb{Cell or Molecular Dysfunction}.}
}
\value{
The expanded set of codes
}
\description{
Expand a set of NCI thesaurus codes related to cell/molecular dysfunctions to include all protein and DNA terms, and all of their ancestors.\cr
For example, \code{C98365} (KRAS G12C, a gene product variation) is expanded to include the respective gene variation(s) (e.g. \code{C98366}), plus all their ancestors (e.g. \code{C135715} KRAS exon 2 mutation, and \code{C98362} KRAS Protein Variant etc)
}
\details{
Relationships described by \code{Gene_Product_Sequence_Variation_Encoded_By_Gene_Mutant} and \code{Gene_Mutant_Encodes_Gene_Product_Sequence_Variation} are added to the specified codes, and then all codes are expanded to include all ancestor terms
}
\seealso{
\code{\link{getNCIt_relationships}} get relationships from NCI thesaurus API.

\code{\link{getAncestors}} get higher level terms
}
