% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAncestors.R
\name{getAncestors}
\alias{getAncestors}
\title{Retrieve Ancestors of NCI Thesaurus Terms}
\usage{
getAncestors(codes, db_connection)
}
\arguments{
\item{codes}{A character vector of NCI Thesaurus concept codes whose ancestors should be retrieved.}

\item{db_connection}{connection to a database that includes the \code{thesaurus_flat} and \code{parents} tables}
}
\value{
A data frame of ancestor concepts, including their codes and preferred terms.
}
\description{
Given a set of codes, finds all ancestor concepts (e.g., parents, grandparents) from the NCI Thesaurus hierarchy.
}
\details{
This function builds and executes a recursive SQL query to retrieve all ancestor nodes for a given list of concept codes.\cr
Ancestors are identified using a parent-child relationship table (\code{NCIt_parents}).\cr
The results are then merged with a pruned thesaurus (\code{NCIt_pruned}) to attach preferred terms (\code{PT}) for easier interpretation.

Internally, the function:
\itemize{
\item Quotes and formats the codes for SQL \code{IN} syntax\cr
\item Uses a recursive CTE (Common Table Expression) to traverse upward through the concept hierarchy\cr
\item Merges results with the pruned thesaurus for human-readable output
}
}
\seealso{
\code{\link[sqldf:sqldf]{sqldf::sqldf()}}, \code{\link[dplyr:select]{dplyr::select()}}, \code{\link[base:merge]{base::merge()}}
}
