% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCodes.R
\name{getCodes}
\alias{getCodes}
\title{Get NCI thesaurus codes from preferred terms}
\usage{
getCodes(PTs, db_connection)
}
\arguments{
\item{PTs}{the preferred term(s)}

\item{db_connection}{connection to a database that includes the thesaurus_flat table}
}
\value{
Matching NCI thesaurus code(s)
}
\description{
Given one or more preferred terms (i.e. first synonyms in NCI thesaurus), look up relevant NCI thesaurus codes.
}
