% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDescendants.R
\name{getDescendants}
\alias{getDescendants}
\title{Retrieve Descendants of NCI Thesaurus Terms}
\usage{
getDescendants(codes, db_connection)
}
\arguments{
\item{codes}{A character vector of NCI Thesaurus concept codes whose descendants should be retrieved.}

\item{db_connection}{connection to a database that includes the \code{thesaurus_flat} and \code{parents} tables}
}
\value{
A character vector of descendant concept codes.
}
\description{
Given a set of codes, finds all descendant concepts (e.g., children, grandchildren) from the NCI Thesaurus hierarchy.
}
\details{
This function builds and executes a recursive SQL query to retrieve all descendant nodes for a given list of concept codes.\cr
Descendants are identified using a parent-child relationship table (\code{NCIt_parents}).\cr
The results are returned as a unique character vector of descendant codes.

Internally, the function:
\itemize{
\item Quotes and formats the input codes for SQL \code{IN} syntax\cr
\item Uses a recursive CTE (Common Table Expression) to traverse downward through the concept hierarchy\cr
\item Returns a de-duplicated vector of descendant concept codes
}
}
\seealso{
\code{\link[sqldf:sqldf]{sqldf::sqldf()}}, \code{\link[dplyr:pull]{dplyr::pull()}}, \code{\link[base:unique]{base::unique()}}
}
