% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNCIt_relationships.R
\name{getNCIt_relationships}
\alias{getNCIt_relationships}
\title{Get NCIt Relationships from NCI Thesaurus API}
\usage{
getNCIt_relationships(code, include_what)
}
\arguments{
\item{code}{A character string representing the NCIt code (concept ID) for which to retrieve relationships.}

\item{include_what}{A character string indicating the type of relationships to include. Valid options are:
\code{roles}, \code{inverseRoles}, \code{associations}, or \code{inverseAssociations}.
This determines which relationship information will be returned for the specified code.}
}
\value{
A data frame containing the specified relationships for the NCIt code. If no relationships are found, it returns \code{NULL}.
}
\description{
This function queries the NCI Thesaurus to retrieve relationships based on a given code and inclusion criteria.
It fetches information from the National Cancer Institute's EVS API for a given concept and returns the specified relationships.
}
\details{
This function constructs a request to the NCI EVS API endpoint to search for the specified NCIt code and includes the relationships
based on the \code{include_what} parameter. The result is parsed into a tibble (data frame), with each relationship type presented
in separate columns.

The following values are valid for \code{include_what}:
\itemize{
\item \code{roles}: Returns roles associated with the concept.
\item \code{inverseRoles}: Returns inverse roles related to the concept.
\item \code{associations}: Returns associations linked to the concept.
\item \code{inverseAssociations}: Returns inverse associations for the concept.
}
}
\seealso{
\url{https://evsexplore.semantics.cancer.gov/evsexplore/evsapi}
}
