% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadThesaurus.R
\name{loadThesaurus}
\alias{loadThesaurus}
\title{Download and Extract the Latest NCI Thesaurus}
\usage{
loadThesaurus()
}
\value{
Path to the extracted NCI Thesaurus flat file.
}
\description{
The latest version of the NCI thesaurus is downloaded and unzipped to the data/raw subfolder.\cr
The NCI Thesaurus is a reference terminology covering a broad range of topics relevant to cancer and biomedical research.
}
\details{
Thesaurus downloaded from \verb{https://evs.nci.nih.gov/ftp1/NCI_Thesaurus/Thesaurus.FLAT.zip} - this should always be the latest version.\cr
}
