% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processThesaurus.R
\name{processThesaurus}
\alias{processThesaurus}
\title{Process the downloaded NCI Thesaurus}
\usage{
processThesaurus(thesaurus_file)
}
\arguments{
\item{thesaurus_file}{The path to the downloaded (raw) version of the NCI Thesaurus.}
}
\value{
Returns a list containing three dataframes:
\itemize{
\item \code{thesaurus}: The processed NCI Thesaurus data with relevant fields extracted and
formatted for analysis, including codes, preferred terms (PT), parents, synonyms,
and semantic types.
\item \code{parents}: A table mapping codes to their parent codes to facilitate hierarchical
analysis of the thesaurus entries.
\item \code{synonyms}: A table of synonyms for each code, expanded from the compressed format
in the source file to aid in text matching and lookup tasks.
}
}
\description{
This function processes the NCI Thesaurus from a flat file format, applying a series of transformations to:
filter obsolete concepts,
extract key information,
prepare the data for analysis.

The transformations include normalising synonyms, identifying solid neoplasm children, and handling generic entities and synonyms.
}
\details{
The NCI Thesaurus is processed through several steps to make it suitable for analysis:
\itemize{
\item Column names are set.
\item A new column 'PT' (preferred term) is added.
\item The first synonym is used as the PT for each entity.
\item Obsolete and retired concepts are removed.
\item New parent-child relationships are added, e.g. between C9292 Solid Neoplasm and various solid tumour types.
\item Generic entities and synonyms that could lead to ambiguous matches are removed.
\item New synonyms are added where appropriate.
\item Parent-child relationships and synonyms are extracted into separate tables for convenient reuse in the pipeline.
}
}
