% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveThesaurus.R
\name{saveThesaurus}
\alias{saveThesaurus}
\title{Save annotated trial data to SQLite database}
\usage{
saveThesaurus(thesaurus_processed, db_connection)
}
\arguments{
\item{thesaurus_processed}{A list of dataframes containing processed information from NCI thesaurus.}

\item{db_connection}{connection to a SQLite database}
}
\value{
Returns the relative path to the populated SQLite database
}
\description{
This is the final step in the trial annotation pipeline.\cr
Data are saved to a local SQLite database for ingestion by user interfaces.
}
\details{
list-columns are flattened into pipe-delimited strings
}
